+++ source/source_base/CMakeLists.txt
@@ -36,7 +36,7 @@ add_library(
     mathzone_add1.cpp
     matrix.cpp
     matrix3.cpp
-    record_memory.cpp
+    memory_recorder.cpp
     mymath.cpp
     opt_CG.cpp
     opt_DCsrch.cpp
