+++ source/source_base/module_device/memory_op.cpp
@@ -1,6 +1,6 @@
 #include "memory_op.h"
 
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/tool_threading.h"
 #ifdef __DSP
 #include "source_base/kernels/dsp/dsp_connector.h"
