+++ source/source_base/test/memory_test.cpp
@@ -26,7 +26,7 @@ namespace GlobalV
  */
 
 #define private public
-#include "../record_memory.h"
+#include "../memory_recorder.h"
 #undef private
 
 class MemoryTest : public testing::Test
