+++ source/source_basis/module_ao/test/CMakeLists.txt
@@ -6,7 +6,7 @@ list(APPEND depend_files
   ../../../source_base/math_polyint.cpp
   ../../../source_base/math_ylmreal.cpp
   ../../../source_base/ylm.cpp
-  ../../../source_base/record_memory.cpp
+  ../../../source_base/memory_recorder.cpp
   ../../../source_base/module_external/blas_connector_base.cpp
   ../../../source_base/module_external/blas_connector_vector.cpp
   ../../../source_base/module_external/blas_connector_matrix.cpp
