+++ source/source_estate/module_charge/charge.cpp
@@ -22,7 +22,7 @@
 #include "source_base/global_variable.h"
 #include "source_base/libm/libm.h"
 #include "source_base/math_integral.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/parallel_reduce.h"
 #include "source_base/timer.h"
 #include "source_base/tool_threading.h"
