+++ source/source_estate/module_dm/density_matrix.cpp
@@ -2,7 +2,7 @@
 
 #include "source_io/module_parameter/parameter.h"
 #include "source_base/libm/libm.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/timer.h"
 #include "source_base/tool_title.h"
 #include "source_base/tool_quit.h"
