+++ source/source_hamilt/module_xc/test/CMakeLists.txt
@@ -37,7 +37,7 @@ AddTest(
     ../xc_funct_corr_gga.cpp ../xc_funct_corr_lda.cpp ../xc_funct_exch_gga.cpp
     ../xc_funct_exch_lda.cpp ../xc_funct_hcth.cpp
     ../../../source_base/matrix.cpp
-    ../../../source_base/record_memory.cpp
+    ../../../source_base/memory_recorder.cpp
     ../../../source_base/libm/branred.cpp
     ../../../source_base/libm/sincos.cpp
     ../../../source_base/module_external/blas_connector_base.cpp ../../../source_base/module_external/blas_connector_vector.cpp ../../../source_base/module_external/blas_connector_matrix.cpp
@@ -75,7 +75,7 @@ AddTest(
     ../xc_functional_libxc_tools.cpp
     ../../../source_base/module_external/blas_connector_base.cpp ../../../source_base/module_external/blas_connector_vector.cpp ../../../source_base/module_external/blas_connector_matrix.cpp
     ../../../source_base/matrix.cpp
-    ../../../source_base/record_memory.cpp
+    ../../../source_base/memory_recorder.cpp
     ../../../source_base/timer.cpp
     ../../../source_base/libm/branred.cpp
     ../../../source_base/libm/sincos.cpp
