+++ source/source_hsolver/hsolver_lcao.cpp
@@ -27,7 +27,7 @@
 #include "source_base/global_variable.h"
 #include "source_base/module_device/device.h"
 #include "source_estate/elecstate_tools.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/timer.h"
 #include "source_estate/elecstate_lcao.h"
 #include "source_estate/module_dm/cal_dm_psi.h"
