+++ source/source_hsolver/parallel_k2d.cpp
@@ -3,7 +3,7 @@
 #include "source_base/parallel_global.h"
 #include "source_base/module_external/scalapack_connector.h"
 #include "source_base/timer.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 
 template <typename TK>
 void Parallel_K2D<TK>::set_para_env(int nks,
