+++ source/source_io/test/CMakeLists.txt
@@ -73,7 +73,7 @@ AddTest(
 	../../source_base/parallel_common.cpp
 	../../source_base/parallel_global.cpp
   ../../source_base/parallel_comm.cpp
-  ../../source_base/tool_quit.cpp ../../source_base/global_file.cpp ../../source_base/global_function.cpp ../../source_base/record_memory.cpp ../../source_base/timer.cpp
+  ../../source_base/tool_quit.cpp ../../source_base/global_file.cpp ../../source_base/global_function.cpp ../../source_base/memory_recorder.cpp ../../source_base/timer.cpp
 )
 
 AddTest(
