+++ source/source_lcao/FORCE_gamma.cpp
@@ -1,5 +1,5 @@
 #include "FORCE.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/parallel_reduce.h"
 #include "source_base/timer.h"
 #include "source_cell/module_neighbor/sltk_grid_driver.h"
