+++ source/source_lcao/edm.cpp
@@ -1,7 +1,7 @@
 #include "FORCE.h"
 #include "source_estate/module_dm/cal_dm_psi.h"
 #include "source_estate/elecstate_lcao.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_io/module_parameter/parameter.h"
 template<>
 elecstate::DensityMatrix<double, double> Force_LCAO<double>::cal_edm(const elecstate::ElecState* pelec,
