+++ source/source_lcao/hamilt_lcao.cpp
@@ -1,7 +1,7 @@
 #include "source_lcao/hamilt_lcao.h"
 
 #include "source_base/global_variable.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/timer.h"
 #include "source_lcao/module_dftu/dftu.h"
 #include "source_io/module_parameter/parameter.h"
