+++ source/source_lcao/module_deepks/test/Makefile.Objects
@@ -26,7 +26,7 @@ math_sphbes.o\
 math_polyint.o\
 math_ylmreal.o\
 ylm.o\
-record_memory.o\
+memory_recorder.o\
 matrix3.o\
 matrix.o\
 intarray.o\
