+++ source/source_lcao/module_dftu/dftu.cpp
@@ -4,7 +4,7 @@
 #include "source_base/constants.h"
 #include "source_base/global_function.h"
 #include "source_base/inverse_matrix.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/timer.h"
 #include "source_estate/magnetism.h"
 #include "source_estate/module_charge/charge.h"
