+++ source/source_lcao/module_gint/gint_info.cpp
@@ -4,7 +4,7 @@
 #include "source_base/timer.h"
 #include "gint_info.h"
 #include "gint_type.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 
 namespace ModuleGint
 {
