+++ source/source_lcao/module_operator_lcao/dspin_lcao.cpp
@@ -1,7 +1,7 @@
 #include "dspin_lcao.h"
 #include "source_lcao/module_deltaspin/spin_constrain.h"
 #include "source_base/timer.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/tool_title.h"
 #include "source_base/parallel_reduce.h"
 #include "source_io/module_parameter/parameter.h"
