+++ source/source_psi/psi_prepare.cpp
@@ -1,7 +1,7 @@
 #include "psi_prepare.h"
 
 #include "source_base/macros.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/parallel_device.h"
 #include "source_base/parallel_global.h"
 #include "source_base/timer.h"
