+++ source/source_pw/module_pwdft/structure_factor.cpp
@@ -4,7 +4,7 @@
 #include "structure_factor.h"
 #include "source_base/constants.h"
 #include "source_base/math_bspline.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/timer.h"
 #include "source_base/libm/libm.h"
