+++ source/source_pw/module_stodft/sto_elecond.cpp
@@ -2,7 +2,7 @@
 
 #include "source_base/complexmatrix.h"
 #include "source_base/constants.h"
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_base/module_container/ATen/tensor.h"
 #include "source_base/parallel_device.h"
 #include "source_base/parallel_reduce.h"
