+++ source/source_pw/module_stodft/sto_wf.cpp
@@ -1,7 +1,7 @@
 #include "sto_wf.h"
 #include "source_base/parallel_comm.h" // use POOL_WORLD
 
-#include "source_base/record_memory.h"
+#include "source_base/memory_recorder.h"
 #include "source_io/module_parameter/parameter.h"
 
 #include <cassert>
